Fix some bad merges
